seccomp: don't commit to an api just yet
I'm not sure that I want to be married (to this layout) just yet.
